+++ system/i18n/bn.php
@@ -1227,4 +1227,7 @@ $_L = [
     'Current plan' => 'Current plan',
     'Plan' => 'Plan',
     'Login as' => 'Login as',
+    'Modules' => 'Modules',
+    'Module' => 'Module',
+    'Enabled' => 'Enabled',
 ];